feat: support copying code snippets
