+++ packages/home/src/data.ts
@@ -1,12 +1,12 @@
 export const COMPONENTS: { option: string; label: string; type: string }[] = [
   { option: 'intro', label: 'Introduction', type: 'introduction' },
-  { option: 'changelog', label: 'Changelog', type: 'introduction' },
+  // { option: 'changelog', label: 'Changelog', type: 'introduction' },
   { option: 'colors', label: 'Colors', type: 'styles' },
-  { option: 'shadows', label: 'Shadows', type: 'styles' },
+  // { option: 'shadows', label: 'Shadows', type: 'styles' },
   { option: 'icons', label: 'Icons', type: 'icons' },
   { option: 'simple-grid', label: 'SimpleGrid', type: 'layout' },
   { option: 'heading', label: 'Heading', type: 'typography' },
-  { option: 'menu', label: 'Menu', type: 'overlay' },
+  // { option: 'menu', label: 'Menu', type: 'overlay' },
   { option: 'button', label: 'Button', type: 'form' },
   { option: 'icon-button', label: 'IconButton', type: 'form' },
   { option: 'radio', label: 'Radio', type: 'form' },
@@ -21,9 +21,9 @@ export const COMPONENTS: { option: string; label: string; type: string }[] = [
   { option: 'avatar', label: 'Avatar', type: 'data display' },
   { option: 'notice', label: 'Notice', type: 'feedback' },
   { option: 'time-picker', label: 'TimePicker', type: 'form' },
-  { option: 'date-picker', label: 'DatePicker', type: 'form' },
+  // { option: 'date-picker', label: 'DatePicker', type: 'form' },
   { option: 'description-list', label: 'DescriptionList', type: 'data display' },
-  { option: 'table', label: 'Table', type: 'data display' },
+  // { option: 'table', label: 'Table', type: 'data display' },
   { option: 'demo1', label: 'Demo: Todo', type: 'demos' },
   { option: 'demo2', label: 'Demo: Issue', type: 'demos' },
 ];
