+++ packages/home/src/docs/date-picker/format.tsx
@@ -1,30 +1,29 @@
-import { DatePicker, Heading, Tabs } from 'fidget-ui';
+import { Heading, DatePicker, Tabs } from 'fidget-ui';
 import { Case } from '../../components/case';
 import { Code } from '../typography';
 
-export default function DatePickerUsage() {
+export default function DatePickerFormat() {
   return (
     <>
-      <Heading as="h5">Usage</Heading>
+      <Heading as="h5">Format</Heading>
 
       <Tabs
+        id="date-picker/format"
         overflow="visible"
-        id="date-picker/usage"
         items={[
           {
             key: 'preview',
             tab: 'Preview',
             children: (
-              <Case height={240} overflow="visible">
-                <DatePicker id="v3k5" />
+              <Case height={240}>
+                <DatePicker id="k843" format={'DD/MM/YYYY'} />
               </Case>
             ),
-            overflow: 'visible',
           },
           {
             key: 'code',
             tab: 'Code',
-            children: <Code height={240}>{`<DatePicker id='v3k5' />`}</Code>,
+            children: <Code height={240}>{`<DatePicker id="k843" format={'DD/MM/YYYY'} />`}</Code>,
           },
         ]}
       />
