+++ packages/home/src/docs/notice/index.tsx
@@ -1,0 +1,24 @@
+import { Divider, Heading } from 'fidget-ui';
+import { docStyle } from '../styles';
+import { P } from '../typography';
+import NoticeUsage from './usage';
+import NoticeTitleDescription from './title-description';
+import NoticeStatus from './status';
+import NoticeIcon from './icon';
+
+const { AutoLayout } = figma.widget;
+
+export function NoticeDoc() {
+  return (
+    <AutoLayout {...docStyle} name="Notice Doc">
+      <Heading as="h1">Notice</Heading>
+      <P>Use Notice components to communicate important messages.</P>
+      <Divider />
+
+      <NoticeUsage />
+      <NoticeTitleDescription />
+      <NoticeStatus />
+      <NoticeIcon />
+    </AutoLayout>
+  );
+}
